divided test module into individual functions as it is likely to become quite large
